+++ apps/block_scout_web/lib/block_scout_web/controllers/tokens/instance/transfer_controller.ex
@@ -12,9 +12,10 @@ defmodule BlockScoutWeb.Tokens.Instance.TransferController do
   {:ok, burn_address_hash} = Chain.string_to_address_hash("0x0000000000000000000000000000000000000000")
   @burn_address_hash burn_address_hash
 
-  def index(conn, %{"token_id" => token_address_hash, "instance_id" => token_id, "type" => "JSON"} = params) do
+  def index(conn, %{"token_id" => token_address_hash, "instance_id" => token_id_str, "type" => "JSON"} = params) do
     with {:ok, hash} <- Chain.string_to_address_hash(token_address_hash),
          {:ok, token} <- Chain.token_from_address_hash(hash),
+         {token_id, ""} <- Integer.parse(token_id_str),
          token_transfers <-
            Chain.fetch_token_transfers_from_token_hash_and_token_id(hash, token_id, paging_options(params)) do
       {token_transfers_paginated, next_page} = split_list_by_page(token_transfers)
@@ -53,21 +54,16 @@ defmodule BlockScoutWeb.Tokens.Instance.TransferController do
     end
   end
 
-  def index(conn, %{"token_id" => token_address_hash, "instance_id" => token_id}) do
+  def index(conn, %{"token_id" => token_address_hash, "instance_id" => token_id_str}) do
     options = [necessity_by_association: %{[contract_address: :smart_contract] => :optional}]
 
     with {:ok, hash} <- Chain.string_to_address_hash(token_address_hash),
          {:ok, token} <- Chain.token_from_address_hash(hash, options),
-         {:ok, token_instance} <-
-           Chain.erc721_or_erc1155_token_instance_from_token_id_and_token_address(token_id, hash) do
-      render(
-        conn,
-        "index.html",
-        token_instance: %{instance: token_instance, token_id: Decimal.new(token_id)},
-        current_path: Controller.current_full_path(conn),
-        token: Market.add_price(token),
-        total_token_transfers: Chain.count_token_transfers_from_token_hash_and_token_id(hash, token_id)
-      )
+         {token_id, ""} <- Integer.parse(token_id_str) do
+      case Chain.erc721_or_erc1155_token_instance_from_token_id_and_token_address(token_id, hash) do
+        {:ok, token_instance} -> render(conn, token_instance, hash, token_id, token)
+        {:error, :not_found} -> render(conn, nil, hash, token_id, token)
+      end
     else
       _ ->
         not_found(conn)
@@ -76,5 +72,16 @@ defmodule BlockScoutWeb.Tokens.Instance.TransferController do
 
   def index(conn, _) do
     not_found(conn)
+  end
+
+  defp render(conn, token_instance, hash, token_id, token) do
+    render(
+      conn,
+      "index.html",
+      token_instance: %{instance: token_instance, token_id: Decimal.new(to_string(token_id))},
+      current_path: Controller.current_full_path(conn),
+      token: Market.add_price(token),
+      total_token_transfers: Chain.count_token_transfers_from_token_hash_and_token_id(hash, token_id)
+    )
   end
 end