+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 
 ### Chore
 
+- [#7097](https://github.com/blockscout/blockscout/pull/7097) - Force display token instance page
 - [#7072](https://github.com/blockscout/blockscout/pull/7072) - Add a separate docker compose for geth with clique consensus
 - [#7056](https://github.com/blockscout/blockscout/pull/7056) - Add path_helper in interact.js
 - [#7040](https://github.com/blockscout/blockscout/pull/7040) - Use alias BlockScoutWeb.Cldr.Number